fixed due to review

+++ app/controllers/accounts_controller.rb
@@ -1,17 +1,17 @@
 class AccountsController < ApplicationController
   def update
-    authorize
+    authorize current_user, policy_class: AddressPolicy
     change_password = ChangePasswordService.new(current_user, permitted_params).call
     if change_password
       flash[:success] = t('flash.change_password_success')
     else
       flash[:error] = t('flash.change_password_error')
     end
-    bypass_sign_in
+    bypass
   end
 
   def destroy
-    authorize
+    authorize current_user, policy_class: AddressPolicy
     if current_user.destroy
       flash[:success] = t('flash.destroy_success')
     else
@@ -22,11 +22,7 @@ class AccountsController < ApplicationController
 
   private
 
-  def authorize
-    authorize current_user, policy_class: AddressPolicy
-  end
-
-  def bypass_sign_in
+  def bypass
     bypass_sign_in(current_user)
     redirect_to edit_address_path(current_user)
   end
